ajax DoS and basic input UI Template
