+++ src/degreesOfSeparation/asset/loader.css
@@ -1,0 +1,4 @@
+<style type='text/css'>
+@-webkit-keyframes uil-default-anim { 0% { opacity: 1} 100% {opacity: 0} }
+@keyframes uil-default-anim { 0% { opacity: 1} 100% {opacity: 0} }
+.uil-default-css > div:nth-of-type(1){-webkit-animation: uil-default-anim 1s linear infinite;animation: uil-default-anim 1s linear infinite;-webkit-animation-delay: -0.5s;animation-delay: -0.5s;}.uil-default-css { position: relative;background:none;width:200px;height:200px;}.uil-default-css > div:nth-of-type(2){-webkit-animation: uil-default-anim 1s linear infinite;animation: uil-default-anim 1s linear infinite;-webkit-animation-delay: -0.4s;animation-delay: -0.4s;}.uil-default-css { position: relative;background:none;width:200px;height:200px;}.uil-default-css > div:nth-of-type(3){-webkit-animation: uil-default-anim 1s linear infinite;animation: uil-default-anim 1s linear infinite;-webkit-animation-delay: -0.3s;animation-delay: -0.3s;}.uil-default-css { position: relative;background:none;width:200px;height:200px;}.uil-default-css > div:nth-of-type(4){-webkit-animation: uil-default-anim 1s linear infinite;animation: uil-default-anim 1s linear infinite;-webkit-animation-delay: -0.2s;animation-delay: -0.2s;}.uil-default-css { position: relative;background:none;width:200px;height:200px;}.uil-default-css > div:nth-of-type(5){-webkit-animation: uil-default-anim 1s linear infinite;animation: uil-default-anim 1s linear infinite;-webkit-animation-delay: -0.09999999999999998s;animation-delay: -0.09999999999999998s;}.uil-default-css { position: relative;background:none;width:200px;height:200px;}.uil-default-css > div:nth-of-type(6){-webkit-animation: uil-default-anim 1s linear infinite;animation: uil-default-anim 1s linear infinite;-webkit-animation-delay: 0s;animation-delay: 0s;}.uil-default-css { position: relative;background:none;width:200px;height:200px;}.uil-default-css > div:nth-of-type(7){-webkit-animation: uil-default-anim 1s linear infinite;animation: uil-default-anim 1s linear infinite;-webkit-animation-delay: 0.09999999999999998s;animation-delay: 0.09999999999999998s;}.uil-default-css { position: relative;background:none;width:200px;height:200px;}.uil-default-css > div:nth-of-type(8){-webkit-animation: uil-default-anim 1s linear infinite;animation: uil-default-anim 1s linear infinite;-webkit-animation-delay: 0.19999999999999996s;animation-delay: 0.19999999999999996s;}.uil-default-css { position: relative;background:none;width:200px;height:200px;}.uil-default-css > div:nth-of-type(9){-webkit-animation: uil-default-anim 1s linear infinite;animation: uil-default-anim 1s linear infinite;-webkit-animation-delay: 0.30000000000000004s;animation-delay: 0.30000000000000004s;}.uil-default-css { position: relative;background:none;width:200px;height:200px;}.uil-default-css > div:nth-of-type(10){-webkit-animation: uil-default-anim 1s linear infinite;animation: uil-default-anim 1s linear infinite;-webkit-animation-delay: 0.4s;animation-delay: 0.4s;}.uil-default-css { position: relative;background:none;width:200px;height:200px;}</style><div class='uil-default-css' style='transform:scale(0.6);'><div style='top:80px;left:95px;width:10px;height:40px;background:#000000;-webkit-transform:rotate(0deg) translate(0,-40px);transform:rotate(0deg) translate(0,-40px);border-radius:16px;position:absolute;'></div><div style='top:80px;left:95px;width:10px;height:40px;background:#000000;-webkit-transform:rotate(36deg) translate(0,-40px);transform:rotate(36deg) translate(0,-40px);border-radius:16px;position:absolute;'></div><div style='top:80px;left:95px;width:10px;height:40px;background:#000000;-webkit-transform:rotate(72deg) translate(0,-40px);transform:rotate(72deg) translate(0,-40px);border-radius:16px;position:absolute;'></div><div style='top:80px;left:95px;width:10px;height:40px;background:#000000;-webkit-transform:rotate(108deg) translate(0,-40px);transform:rotate(108deg) translate(0,-40px);border-radius:16px;position:absolute;'></div><div style='top:80px;left:95px;width:10px;height:40px;background:#000000;-webkit-transform:rotate(144deg) translate(0,-40px);transform:rotate(144deg) translate(0,-40px);border-radius:16px;position:absolute;'></div><div style='top:80px;left:95px;width:10px;height:40px;background:#000000;-webkit-transform:rotate(180deg) translate(0,-40px);transform:rotate(180deg) translate(0,-40px);border-radius:16px;position:absolute;'></div><div style='top:80px;left:95px;width:10px;height:40px;background:#000000;-webkit-transform:rotate(216deg) translate(0,-40px);transform:rotate(216deg) translate(0,-40px);border-radius:16px;position:absolute;'></div><div style='top:80px;left:95px;width:10px;height:40px;background:#000000;-webkit-transform:rotate(252deg) translate(0,-40px);transform:rotate(252deg) translate(0,-40px);border-radius:16px;position:absolute;'></div><div style='top:80px;left:95px;width:10px;height:40px;background:#000000;-webkit-transform:rotate(288deg) translate(0,-40px);transform:rotate(288deg) translate(0,-40px);border-radius:16px;position:absolute;'></div><div style='top:80px;left:95px;width:10px;height:40px;background:#000000;-webkit-transform:rotate(324deg) translate(0,-40px);transform:rotate(324deg) translate(0,-40px);border-radius:16px;position:absolute;'></div></div>
