+++ src/degreesOfSeparation/html/html.go
@@ -1,4 +1,4 @@
-package main
+package html
 
 import (
 	"html/template"
